Move serialize_async into its own module.
