+++ src/lib.rs
@@ -85,6 +85,7 @@ pub mod message;
 pub mod primitive_list;
 pub mod private;
 pub mod serialize;
+pub mod serialize_async;
 pub mod serialize_packed;
 pub mod struct_list;
 pub mod text;
