+++ src/serialize.rs
@@ -22,10 +22,11 @@
 //! Reading and writing of messages using the
 //! [standard stream framing](https://capnproto.org/encoding.html#serialization-over-a-stream).
 
-use std::io::{self, Read, Write};
+use std::io::{Read, Write};
 
 use message;
-use {Error, Result, Word, OutputSegments};
+use util::read_exact;
+use {Error, Result, Word};
 
 use byteorder::{ByteOrder, LittleEndian};
 
@@ -51,24 +52,21 @@ pub fn read_message_from_words<'a>(slice: &'a [Word],
                                    options: message::ReaderOptions)
                                    -> Result<message::Reader<SliceSegments<'a>>>
 {
-    let bytes = ::Word::words_to_bytes(slice);
-    let (segment_count, first_segment_len) = try!(parse_segment_table_first(&bytes[..8]));
-    let mut segment_slices = Vec::with_capacity(segment_count as usize);
-    segment_slices.push((0, first_segment_len));
-
-    let bufsize = 4 * (segment_count as usize & !1);
-    let (_, offsets) =
-        try!(parse_segment_table_rest(
-            &options, segment_count, first_segment_len,
-            &bytes[8..(8 + bufsize)]));
-
-    let words = ::Word::bytes_to_words(&bytes[(8 + bufsize)..]);
-    Ok(message::Reader::new(SliceSegments { words: words, segment_slices: offsets }, options))
+    let mut bytes = ::Word::words_to_bytes(slice);
+    let (num_words, offsets) = try!(read_segment_table(&mut bytes, options));
+    let words = ::Word::bytes_to_words(bytes);
+    if num_words != words.len() {
+        Err(Error::failed(
+            format!("Wrong number of words. Header claimed {} words, but message has {} words",
+                    num_words, words.len())))
+    } else {
+        Ok(message::Reader::new(SliceSegments { words: words, segment_slices: offsets }, options))
+    }
 }
 
 pub struct OwnedSegments {
-    segment_slices: Vec<(usize, usize)>,
-    owned_space: Vec<Word>,
+    segment_slices : Vec<(usize, usize)>,
+    owned_space : Vec<Word>,
 }
 
 impl ::message::ReaderSegments for OwnedSegments {
@@ -82,274 +80,63 @@ impl ::message::ReaderSegments for OwnedSegments {
     }
 }
 
-/// Reads bytes from `read` into `buf` until either `buf` is full, or the read
-/// would block. Returns the number of bytes read.
-fn async_read_all<R>(read: &mut R, buf: &mut [u8]) -> io::Result<usize> where R: Read {
-    let mut idx = 0;
-    while idx < buf.len() {
-        let slice = &mut buf[idx..];
-        match read.read(slice) {
-            Ok(n) if n == 0 => return Err(io::Error::new(io::ErrorKind::Other, "Premature EOF")),
-            Ok(n) => idx += n,
-            Err(ref e) if e.kind() == io::ErrorKind::WouldBlock => break,
-            Err(ref e) if e.kind() == io::ErrorKind::Interrupted => (),
-            Err(e) => return Err(e),
-        }
-    }
-    return Ok(idx)
-}
-
-/// Writes bytes from `buf` into `write` until either all bytes are written, or
-/// the write would block. Returns the number of bytes written.
-fn async_write_all<W>(write: &mut W, buf: &[u8]) -> io::Result<usize> where W: Write {
-    let mut idx = 0;
-    while idx < buf.len() {
-        let slice = &buf[idx..];
-        match write.write(slice) {
-            Ok(n) if n == 0 => return Err(io::Error::new(io::ErrorKind::WriteZero,
-                                                         "failed to write whole buffer")),
-            Ok(n) => idx += n,
-            Err(ref e) if e.kind() == io::ErrorKind::WouldBlock => break,
-            Err(ref e) if e.kind() == io::ErrorKind::Interrupted => (),
-            Err(e) => return Err(e),
-        }
-    }
-    return Ok(idx)
-}
-
-pub enum Async<T> {
-    Ready(T),
-    NotReady,
-}
-
-/// An in-progress read operation.
-#[derive(Debug)]
-pub struct AsyncRead<R> where R: ::std::io::Read {
-    state: ReadState<R>,
-}
-
-#[derive(Debug)]
-enum ReadState<R> where R: ::std::io::Read {
-    Reading {
-        read: R,
-        options: message::ReaderOptions,
-        inner: InnerReadState,
-    },
-
-    Empty,
-}
-
-#[derive(Debug)]
-pub enum InnerReadState {
-    /// Reading the message would block while trying to read the first word (the
-    /// segment count, and the first segment's length).
-    SegmentTableFirst {
-        /// The buffer being read into.
-        buf: [u8; 8],
-        /// The number of bytes read before being blocked.
-        idx: usize,
-    },
-
-    /// Reading the message would block while trying to read the rest of the segment table.
-    SegmentTableRest {
-        segment_count: u32,
-        first_segment_length: u32,
-        /// The buffer being read into.
-        segment_size_buf: Vec<u8>,
-        /// The number of bytes read before being blocked.
-        idx: usize,
-    },
-
-    /// Reading the message would block while trying to read the segments.
-    Segments {
-        segment_slices: Vec<(usize, usize)>,
-        /// The segment buffer.
-        owned_space: Vec<Word>,
-        /// The number of bytes read into `owned_space` before being blocked.
-        idx: usize,
-    },
-}
-
-/// Reads a Cap'n Proto serialized message from a stream with the provided options.
+/// Reads a serialized message from a stream with the provided options.
 ///
-/// For optimal performance, `reader` should be a buffered reader type.
-pub fn read_message<R>(reader: &mut R, options: message::ReaderOptions) -> Result<message::Reader<OwnedSegments>>
-    where R: Read
-{
-    let mut future = read_message_async(reader, options);
-    match future.poll() {
-        Ok(Async::Ready((_, m))) => Ok(m),
-        Ok(Async::NotReady) => Err(::Error::failed("message read would block".into())),
-        Err(e) => Err(e)
-    }
+/// For optimal performance, `read` should be a buffered reader type.
+pub fn read_message<R>(read: &mut R, options: message::ReaderOptions) -> Result<message::Reader<OwnedSegments>>
+where R: Read {
+    let (total_words, segment_slices) = try!(read_segment_table(read, options));
+    read_segments(read, total_words, segment_slices, options)
 }
 
-/// Asynchronously reads a message from `reader`.
-pub fn read_message_async<R>(reader: R, options: message::ReaderOptions) -> AsyncRead<R>
-    where R: Read
-{
-    AsyncRead {
-        state: ReadState::Reading {
-            read: reader,
-            options: options,
-            inner: InnerReadState::SegmentTableFirst { buf: [0; 8], idx: 0 },
-        }
-    }
-}
-
-impl InnerReadState {
-    fn read_helper<R>(&mut self, read: &mut R, options: &message::ReaderOptions)
-                      -> Result<Async<(Vec<Word>, Vec<(usize, usize)>)>>
-        where R: ::std::io::Read
-    {
-        loop {
-            let next_state = match *self {
-                InnerReadState::SegmentTableFirst { ref mut buf, ref mut idx } => {
-                    *idx += try!(async_read_all(read, &mut buf[*idx..]));
-                    if *idx < buf.len() {
-                        return Ok(Async::NotReady)
-                    } else {
-                        let (segment_count, first_segment_length) = try!(parse_segment_table_first(buf));
-                        if segment_count == 1 {
-                            InnerReadState::Segments {
-                                segment_slices: vec![(0, first_segment_length as usize)],
-                                owned_space: ::Word::allocate_zeroed_vec(first_segment_length as usize),
-                                idx: 0,
-                            }
-                        } else {
-                            InnerReadState::SegmentTableRest {
-                                segment_count: segment_count,
-                                first_segment_length: first_segment_length,
-                                segment_size_buf: vec![0u8; 4 * (segment_count as usize & !1)],
-                                idx: 0,
-                            }
-                        }
-                    }
-                }
-
-                InnerReadState::SegmentTableRest {
-                    segment_count,
-                    first_segment_length,
-                    ref mut segment_size_buf,
-                    ref mut idx,
-                } => {
-                    *idx += try!(async_read_all(read, &mut segment_size_buf[*idx..]));
-                    if *idx < segment_size_buf.len() {
-                        return Ok(Async::NotReady)
-                    } else {
-                        let (word_count, segment_slices) =
-                            try!(parse_segment_table_rest(
-                                options, segment_count, first_segment_length, segment_size_buf));
-                        InnerReadState::Segments {
-                            segment_slices: segment_slices,
-                            owned_space: ::Word::allocate_zeroed_vec(word_count),
-                            idx: 0,
-                        }
-                    }
-                }
-
-                InnerReadState::Segments { ref mut segment_slices, ref mut owned_space, ref mut idx } => {
-                    let len = {
-                        let mut bytes = ::Word::words_to_bytes_mut(owned_space);
-                        *idx += try!(async_read_all(read, &mut bytes[*idx..]));
-                        bytes.len()
-                    };
-                    if *idx < len {
-                        return Ok(Async::NotReady)
-                    } else {
-                        let words = ::std::mem::replace(owned_space, Vec::new());
-                        let slices = ::std::mem::replace(segment_slices, Vec::new());
-                        return Ok(Async::Ready((words, slices)))
-                    }
-                }
-            };
-
-            *self = next_state;
-        }
-    }
-}
-
-impl <R> AsyncRead<R> where R: ::std::io::Read {
-    /// Reads a Cap'n Proto serialized message from `read` with the provided options.
-    pub fn poll(&mut self) -> Result<Async<(R, message::Reader<OwnedSegments>)>> {
-        let (words, slices) = match &mut self.state {
-            &mut ReadState::Empty => {
-                return Err(::Error::failed("tried to read empty ReadState".to_string()))
-            }
-            &mut ReadState::Reading { ref mut read, ref options, ref mut inner } => {
-                match try!(inner.read_helper(read, options)) {
-                    Async::NotReady => return Ok(Async::NotReady),
-                    Async::Ready(v) => v,
-                }
-            }
-        };
-
-        let old_self = ::std::mem::replace(&mut self.state, ReadState::Empty);
-        match old_self {
-            ReadState::Empty => unreachable!(),
-            ReadState::Reading { read, options, ..} => {
-                return Ok(Async::Ready((
-                    read,
-                    message::Reader::new(
-                        OwnedSegments {
-                            segment_slices: slices,
-                            owned_space: words,
-                        },
-                        options))))
-            }
-        }
-    }
-}
-
-#[cfg(feature="futures")]
-impl <R> ::futures::Future for AsyncRead<R> where R: ::std::io::Read {
-    type Item = (R, message::Reader<OwnedSegments>);
-    type Error = Error;
-    fn poll(&mut self) -> ::futures::Poll<(R, message::Reader<OwnedSegments>), Error> {
-        match try!(AsyncRead::poll(self)) {
-            Async::Ready(v) => Ok(::futures::Async::Ready(v)),
-            Async::NotReady => Ok(::futures::Async::NotReady),
-        }
-    }
-}
-
-/// Reads or continues reading the first word of a segment table from `read`.
-///
-/// Returns the segment count and first segment length, or a state if the
-/// read would block.
+/// Reads a segment table from `read` and returns the total number of words across all
+/// segments, as well as the segment offsets.
 ///
 /// The segment table format for streams is defined in the Cap'n Proto
-/// [encoding spec](https://capnproto.org/encoding.html#serialization-over-a-stream)
-fn parse_segment_table_first(buf: &[u8]) -> Result<(u32, u32)>
+/// [encoding spec](https://capnproto.org/encoding.html)
+fn read_segment_table<R>(read: &mut R,
+                         options: message::ReaderOptions)
+                         -> Result<(usize, Vec<(usize, usize)>)>
+    where R: Read
 {
+    let mut buf: [u8; 8] = [0; 8];
+
+    // read the first Word, which contains segment_count and the 1st segment length
+    try!(read_exact(read, &mut buf));
     let segment_count = <LittleEndian as ByteOrder>::read_u32(&buf[0..4])
-                                                   .wrapping_add(1);
+                                                   .wrapping_add(1) as usize;
+
     if segment_count >= 512 {
         return Err(Error::failed(format!("Too many segments: {}", segment_count)))
     } else if segment_count == 0 {
         return Err(Error::failed(format!("Too few segments: {}", segment_count)))
     }
 
-    let first_segment_len = <LittleEndian as ByteOrder>::read_u32(&buf[4..8]);
-    Ok((segment_count, first_segment_len))
-}
+    let mut segment_slices = Vec::with_capacity(segment_count);
+    let mut total_words = <LittleEndian as ByteOrder>::read_u32(&buf[4..8]) as usize;
+    segment_slices.push((0, total_words));
 
-/// Reads or continues reading the remaining words (after the first) of a
-/// segment table from `read`.
-///
-/// Returns the total segment words and segment slices, or a state if the
-/// read would block.
-fn parse_segment_table_rest(options: &message::ReaderOptions,
-                            segment_count: u32,
-                            first_segment_length: u32,
-                            buf: &[u8]) -> Result<(usize, Vec<(usize, usize)>)> {
-    let mut total_words = first_segment_length as usize;
-    let mut segment_slices = vec![(0usize, first_segment_length as usize)];
+    if segment_count > 1 {
+        if segment_count < 4 {
+            try!(read_exact(read, &mut buf));
+            for idx in 0..(segment_count - 1) {
+                let segment_len =
+                    <LittleEndian as ByteOrder>::read_u32(&buf[(idx * 4)..(idx + 1) * 4]) as usize;
 
-    for idx in 0..(segment_count as usize - 1) {
-        let segment_len = <LittleEndian as ByteOrder>::read_u32(&buf[(idx * 4)..(idx * 4 + 4)]) as usize;
-        segment_slices.push((total_words, total_words + segment_len));
-        total_words += segment_len;
+                segment_slices.push((total_words, total_words + segment_len));
+                total_words += segment_len;
+            }
+        } else {
+            let mut segment_sizes = vec![0u8; (segment_count & !1) * 4];
+            try!(read_exact(read, &mut segment_sizes[..]));
+            for idx in 0..(segment_count - 1) {
+                let segment_len =
+                    <LittleEndian as ByteOrder>::read_u32(&segment_sizes[(idx * 4)..(idx + 1) * 4]) as usize;
+
+                segment_slices.push((total_words, total_words + segment_len));
+                total_words += segment_len;
+            }
+        }
     }
 
     // Don't accept a message which the receiver couldn't possibly traverse without hitting the
@@ -364,207 +151,101 @@ fn parse_segment_table_rest(options: &message::ReaderOptions,
     Ok((total_words, segment_slices))
 }
 
-/// An in-progress write operation.
-#[derive(Debug)]
-pub struct AsyncWrite<W, M> where W: ::std::io::Write, M: GetOutputSegments {
-    state: WriteState<W, M>,
-}
-
-#[derive(Debug)]
-enum WriteState<W, M> where W: ::std::io::Write, M: GetOutputSegments, {
-    Writing {
-        writer: W,
-        message: M,
-        inner: InnerWriteState,
-    },
-    Empty,
-}
-
-fn construct_segment_table(segments: &[&[Word]]) -> Vec<u8> {
-    let mut buf = vec![0u8; (((segments.len() + 2) & !1) * 4)];
-    <LittleEndian as ByteOrder>::write_u32(&mut buf[0..4], segments.len() as u32 - 1);
-    for idx in 0..segments.len() {
-        <LittleEndian as ByteOrder>::write_u32(
-            &mut buf[(idx + 1) * 4 .. (idx + 2) * 4],
-            segments[idx].len() as u32);
-    }
-    buf
-}
-
-#[derive(Debug)]
-enum InnerWriteState {
-    /// Writing the message would block while trying to write the segment table.
-    OneWordSegmentTable {
-        buf: [u8; 8],
-        idx: usize,
-    },
-
-    MoreThanOneWordSegmentTable {
-        buf: Vec<u8>,
-        idx: usize,
-    },
-
-    /// Writing the message would block while trying to write the message segments.
-    Segments {
-        /// The next segment to write.
-        segment_idx: usize,
-        /// The byte offset into the next segment to write.
-        idx: usize,
-    },
-}
-
-impl InnerWriteState {
-    fn write_helper<W, M>(&mut self, writer: &mut W, message: &mut M)
-                          -> io::Result<Async<()>>
-        where W: ::std::io::Write, M: GetOutputSegments,
-    {
-        loop {
-            let new_state = match *self {
-                InnerWriteState::OneWordSegmentTable { ref mut buf, ref mut idx } => {
-                    *idx += try!(async_write_all(writer, &buf[*idx..]));
-                    if *idx < 8 {
-                        return Ok(Async::NotReady)
-                    } else {
-                        InnerWriteState::Segments { segment_idx: 0, idx: 0 }
-                    }
-                }
-                InnerWriteState::MoreThanOneWordSegmentTable { ref mut buf, ref mut idx } => {
-                    *idx += try!(async_write_all(writer, &buf[*idx..]));
-                    if *idx < buf.len() {
-                        return Ok(Async::NotReady)
-                    } else {
-                        InnerWriteState::Segments { segment_idx: 0, idx: 0 }
-                    }
-                }
-                InnerWriteState::Segments { ref mut segment_idx, ref mut idx } => {
-                    let segments = &*message.get_output_segments();
-                    while *segment_idx < segments.len() {
-                        let segment = segments[*segment_idx];
-                        let buf = ::Word::words_to_bytes(segment);
-                        *idx += try!(async_write_all(writer, &buf[*idx..]));
-                        if *idx < buf.len() {
-                            return Ok(Async::NotReady)
-                        } else {
-                            *segment_idx += 1;
-                            *idx = 0;
-                        }
-                    }
-
-                    return Ok(Async::Ready(()))
-                }
-            };
-
-            *self = new_state;
-        }
-    }
-}
-
-#[cfg(feature="futures")]
-impl <W, M> ::futures::Future for AsyncWrite<W, M> where W: ::std::io::Write, M: GetOutputSegments {
-    type Item = (W, M);
-    type Error = Error;
-    fn poll(&mut self) -> ::futures::Poll<(W, M), Error> {
-        match try!(AsyncWrite::poll(self)) {
-            Async::Ready(v) => Ok(::futures::Async::Ready(v)),
-            Async::NotReady => Ok(::futures::Async::NotReady),
-        }
-    }
+/// Reads segments from `read`.
+fn read_segments<R>(read: &mut R,
+                    total_words: usize,
+                    segment_slices: Vec<(usize, usize)>,
+                    options: message::ReaderOptions)
+                    -> Result<message::Reader<OwnedSegments>>
+where R: Read {
+    let mut owned_space: Vec<Word> = Word::allocate_zeroed_vec(total_words);
+    try!(read_exact(read, Word::words_to_bytes_mut(&mut owned_space[..])));
+    let segments = OwnedSegments {segment_slices: segment_slices, owned_space: owned_space};
+    Ok(::message::Reader::new(segments, options))
 }
 
 /// Constructs a flat vector containing the entire message.
 pub fn write_message_to_words<A>(message: &message::Builder<A>) -> Vec<Word>
     where A: message::Allocator
 {
-    let word_count = compute_serialized_size_in_words(message);
-    let mut result = ::Word::allocate_zeroed_vec(word_count);
+    flatten_segments(&*message.get_segments_for_output())
+}
+
+fn flatten_segments(segments: &[&[Word]]) -> Vec<Word> {
+    let word_count = compute_serialized_size(&*segments);
+    let table_size = segments.len() / 2 + 1;
+    let mut result = Vec::with_capacity(word_count);
+    for _ in 0..table_size {
+        result.push(Word { raw_content: 0 });
+    }
     {
         let mut bytes = ::Word::words_to_bytes_mut(&mut result[..]);
-        write_message(&mut bytes, message).unwrap();
+        write_segment_table(&mut bytes, &*segments).ok().expect("Failed to write segment table.");
     }
-
+    for segment in &*segments {
+        for idx in 0..segment.len() {
+            result.push(segment[idx]);
+        }
+    }
     result
 }
 
-/// Something that contains segments ready to be written out.
-pub trait GetOutputSegments {
-    fn get_output_segments<'a>(&'a self) -> OutputSegments<'a>;
-}
-
-impl <A> GetOutputSegments for message::Builder<A> where A: message::Allocator {
-    fn get_output_segments<'a>(&'a self) -> OutputSegments<'a> {
-        self.get_segments_for_output()
-    }
-}
-
-impl <'a, A> GetOutputSegments for &'a message::Builder<A> where A: message::Allocator {
-    fn get_output_segments<'b>(&'b self) -> OutputSegments<'b> {
-        self.get_segments_for_output()
-    }
-}
-
-/// Writes the provided message to `writer`.
+/// Writes the provided message to `write`.
 ///
-/// For optimal performance, `writer` should be a buffered writer. `flush` will not be called on
+/// For optimal performance, `write` should be a buffered writer. `flush` will not be called on
 /// the writer.
-pub fn write_message<W, A>(writer: &mut W, message: &message::Builder<A>) -> io::Result<()>
-      where W: Write, A: message::Allocator
-{
-    let mut future = write_message_async(writer, message);
-    match future.poll() {
-        Ok(Async::Ready((_, _))) => Ok(()),
-        Ok(Async::NotReady) => Err(io::Error::new(io::ErrorKind::Other, "message read would block")),
-        Err(e) => Err(e)
-    }
+pub fn write_message<W, A>(write: &mut W, message: &message::Builder<A>) -> ::std::io::Result<()>
+where W: Write, A: message::Allocator {
+    let segments = message.get_segments_for_output();
+    try!(write_segment_table(write, &*segments));
+    write_segments(write, &*segments)
 }
 
-/// Begins an asynchronous write of provided message to `writer`.
-pub fn write_message_async<W, M>(writer: W, message: M) -> AsyncWrite<W, M>
-    where W: ::std::io::Write, M: GetOutputSegments
-{
-    let inner = {
-        let segments = &*message.get_output_segments();
-        if segments.len() == 1 {
-            let mut buf = [0; 8];
-            <LittleEndian as ByteOrder>::write_u32(&mut buf[4..8], segments[0].len() as u32);
-            InnerWriteState::OneWordSegmentTable { buf: buf, idx: 0 }
+/// Writes a segment table to `write`.
+///
+/// `segments` must contain at least one segment.
+fn write_segment_table<W>(write: &mut W, segments: &[&[Word]]) -> ::std::io::Result<()>
+where W: Write {
+    let mut buf: [u8; 8] = [0; 8];
+    let segment_count = segments.len();
+
+    // write the first Word, which contains segment_count and the 1st segment length
+    <LittleEndian as ByteOrder>::write_u32(&mut buf[0..4], segment_count as u32 - 1);
+    <LittleEndian as ByteOrder>::write_u32(&mut buf[4..8], segments[0].len() as u32);
+    try!(write.write_all(&buf));
+
+    if segment_count > 1 {
+        if segment_count < 4 {
+            for idx in 1..segment_count {
+                <LittleEndian as ByteOrder>::write_u32(
+                    &mut buf[(idx - 1) * 4..idx * 4], segments[idx].len() as u32);
+            }
+            if segment_count == 2 {
+                for idx in 4..8 { buf[idx] = 0 }
+            }
+            try!(write.write_all(&buf));
         } else {
-            let buf = construct_segment_table(segments);
-            InnerWriteState::MoreThanOneWordSegmentTable { buf: buf, idx: 0 }
-        }
-    };
-
-    AsyncWrite {
-        state: WriteState::Writing {
-            writer: writer,
-            message: message,
-            inner: inner,
+            let mut buf = vec![0; (segment_count & !1) * 4];
+            for idx in 1..segment_count {
+                <LittleEndian as ByteOrder>::write_u32(
+                    &mut buf[(idx - 1) * 4..idx * 4], segments[idx].len() as u32);
+            }
+            if segment_count % 2 == 0 {
+                for idx in (buf.len() - 4)..(buf.len()) { buf[idx] = 0 }
+            }
+            try!(write.write_all(&buf));
         }
     }
+    Ok(())
 }
 
-impl <W, M> AsyncWrite<W, M> where W: ::std::io::Write, M: GetOutputSegments {
-    pub fn poll(&mut self) -> io::Result<Async<(W, M)>>
-    {
-        match self.state {
-            WriteState::Empty => {
-                return Err(io::Error::new(io::ErrorKind::Other, "tried to poll empty AsyncWrite".to_string()))
-            }
-            WriteState::Writing { ref mut writer, ref mut message, ref mut inner } => {
-                match try!(inner.write_helper(writer, message)) {
-                    Async::NotReady => return Ok(Async::NotReady),
-                    Async::Ready(()) => (),
-                }
-            }
-        };
-
-        let old_self = ::std::mem::replace(&mut self.state, WriteState::Empty);
-        match old_self {
-            WriteState::Empty => unreachable!(),
-            WriteState::Writing { writer, message, ..} => {
-                Ok(Async::Ready((writer, message)))
-            }
-        }
+/// Writes segments to `write`.
+fn write_segments<W>(write: &mut W, segments: &[&[Word]]) -> ::std::io::Result<()>
+where W: Write {
+    for segment in segments {
+        try!(write.write_all(Word::words_to_bytes(segment)));
     }
+    Ok(())
 }
 
 fn compute_serialized_size(segments: &[&[Word]]) -> usize {
@@ -586,52 +267,23 @@ pub fn compute_serialized_size_in_words<A>(message: &::message::Builder<A>) -> u
 #[cfg(test)]
 pub mod test {
 
-    use std::cmp;
-    use std::io::{self, Cursor, Read, Write};
+    use std::io::{Cursor, Write};
 
     use quickcheck::{quickcheck, TestResult};
 
-    use {Word, message, Result, OutputSegments};
+    use {Word};
+    use message;
     use message::ReaderSegments;
-
-    use super::{
-        Async,
-        GetOutputSegments,
-        read_message,
-        read_message_async,
-        read_message_from_words,
-        parse_segment_table_first,
-        parse_segment_table_rest,
-        compute_serialized_size,
-        construct_segment_table,
-        write_message_async,
-    };
-
-    pub fn read_segment_table<R>(read: &mut R,
-                                 options: message::ReaderOptions)
-                                 -> Result<(usize, Vec<(usize, usize)>)>
-    where R: Read {
-        let mut firstbuf = [0; 8];
-        try!(read.read_exact(&mut firstbuf));
-        let (segment_count, first_segment_len) = try!(parse_segment_table_first(&firstbuf[..]));
-
-        let mut rest_buf = vec![0; 4*(segment_count as usize & !1)];
-        try!(read.read_exact(&mut rest_buf));
-
-        parse_segment_table_rest(&options, segment_count, first_segment_len, &rest_buf[..])
-    }
+    use super::{read_message, read_message_from_words, flatten_segments,
+                read_segment_table, write_segment_table, write_segments};
 
     /// Writes segments as if they were a Capnproto message.
-    pub fn write_message_segments<W>(writer: &mut W, segments: &Vec<Vec<Word>>) where W: Write {
+    pub fn write_message_segments<W>(write: &mut W, segments: &Vec<Vec<Word>>) where W: Write {
         let borrowed_segments: &[&[Word]] = &segments.iter()
                                                      .map(|segment| &segment[..])
                                                      .collect::<Vec<_>>()[..];
-
-        let outbuf = construct_segment_table(borrowed_segments);
-        writer.write_all(&outbuf[..]).unwrap();
-        for segment in segments {
-            writer.write_all(Word::words_to_bytes(segment)).unwrap()
-        }
+        write_segment_table(write, borrowed_segments).unwrap();
+        write_segments(write, borrowed_segments).unwrap();
     }
 
     #[test]
@@ -722,35 +374,42 @@ pub mod test {
     }
 
     #[test]
-    fn test_construct_segment_table() {
+    fn test_write_segment_table() {
+
+        let mut buf = vec![];
 
         let segment_0 = [Word::from(0); 0];
         let segment_1 = [Word::from(1); 1];
         let segment_199 = [Word::from(199); 199];
 
-        let buf = construct_segment_table(&[&segment_0]);
+        write_segment_table(&mut buf, &[&segment_0]).unwrap();
         assert_eq!(&[0,0,0,0,  // 1 segments
                      0,0,0,0], // 0 length
                    &buf[..]);
+        buf.clear();
 
-        let buf = construct_segment_table(&[&segment_1]);
+        write_segment_table(&mut buf, &[&segment_1]).unwrap();
         assert_eq!(&[0,0,0,0,  // 1 segments
                      1,0,0,0], // 1 length
                    &buf[..]);
+        buf.clear();
 
-        let buf = construct_segment_table(&[&segment_199]);
+        write_segment_table(&mut buf, &[&segment_199]).unwrap();
         assert_eq!(&[0,0,0,0,    // 1 segments
                      199,0,0,0], // 199 length
                    &buf[..]);
+        buf.clear();
 
-        let buf = construct_segment_table(&[&segment_0, &segment_1]);;
+        write_segment_table(&mut buf, &[&segment_0, &segment_1]).unwrap();
         assert_eq!(&[1,0,0,0,  // 2 segments
                      0,0,0,0,  // 0 length
                      1,0,0,0,  // 1 length
                      0,0,0,0], // padding
                    &buf[..]);
+        buf.clear();
 
-        let buf = construct_segment_table(&[&segment_199, &segment_1, &segment_199, &segment_0]);
+        write_segment_table(&mut buf,
+                            &[&segment_199, &segment_1, &segment_199, &segment_0]).unwrap();
         assert_eq!(&[3,0,0,0,   // 4 segments
                      199,0,0,0, // 199 length
                      1,0,0,0,   // 1 length
@@ -758,9 +417,10 @@ pub mod test {
                      0,0,0,0,   // 0 length
                      0,0,0,0],  // padding
                    &buf[..]);
+        buf.clear();
 
-        let buf = construct_segment_table(
-            &[&segment_199, &segment_1, &segment_199, &segment_0, &segment_1]);
+        write_segment_table(&mut buf,
+                            &[&segment_199, &segment_1, &segment_199, &segment_0, &segment_1]).unwrap();
         assert_eq!(&[4,0,0,0,   // 5 segments
                      199,0,0,0, // 199 length
                      1,0,0,0,   // 1 length
@@ -768,20 +428,7 @@ pub mod test {
                      0,0,0,0,   // 0 length
                      1,0,0,0],  // 1 length
                    &buf[..]);
-    }
-
-    impl GetOutputSegments for Vec<Vec<Word>> {
-        fn get_output_segments<'a>(&'a self) -> OutputSegments<'a> {
-            if self.len() == 0 {
-                OutputSegments::SingleSegment([&[]])
-            } else if self.len() == 1 {
-                OutputSegments::SingleSegment([&self[0][..]])
-            } else {
-                OutputSegments::MultiSegment(self.iter()
-                                             .map(|segment| &segment[..])
-                                             .collect::<Vec<_>>())
-            }
-        }
+        buf.clear();
     }
 
     #[test]
@@ -808,25 +455,10 @@ pub mod test {
     fn check_round_trip_slice_segments() {
         fn round_trip(segments: Vec<Vec<Word>>) -> TestResult {
             if segments.len() == 0 { return TestResult::discard(); }
-            let (word_count, segment_table) = {
-                let borrowed_segments: &[&[Word]] = &segments.iter()
-                    .map(|segment| &segment[..])
-                    .collect::<Vec<_>>()[..];
-
-                (compute_serialized_size(borrowed_segments),
-                 construct_segment_table(borrowed_segments))
-            };
-
-            let mut words = ::Word::allocate_zeroed_vec(word_count);
-            {
-                let mut bytes = ::Word::words_to_bytes_mut(&mut words[..]);
-                bytes.write_all(&segment_table[..]).unwrap();
-                for segment in &segments {
-                    let segment_bytes = ::Word::words_to_bytes(segment);
-                    bytes.write_all(segment_bytes).unwrap();
-                }
-            }
-
+            let borrowed_segments: &[&[Word]] = &segments.iter()
+                                                     .map(|segment| &segment[..])
+                                                     .collect::<Vec<_>>()[..];
+            let words = flatten_segments(&borrowed_segments);
             let message = read_message_from_words(&words[..], message::ReaderOptions::new()).unwrap();
             let result_segments = message.into_segments();
 
@@ -836,124 +468,5 @@ pub mod test {
         }
 
         quickcheck(round_trip as fn(Vec<Vec<Word>>) -> TestResult);
-    }
-
-    /// Wraps a `Read` instance and introduces blocking.
-    struct BlockingRead<R> where R: Read {
-        /// The wrapped reader
-        read: R,
-
-        /// Number of bytes to read before blocking
-        frequency: usize,
-
-        /// Number of bytes read since last blocking
-        idx: usize,
-    }
-
-    impl <R> BlockingRead<R> where R: Read {
-        fn new(read: R, frequency: usize) -> BlockingRead<R> {
-            BlockingRead { read: read, frequency: frequency, idx: 0 }
-        }
-    }
-
-    impl <R> Read for BlockingRead<R> where R: Read {
-        fn read(&mut self, buf: &mut [u8]) -> io::Result<usize> {
-            if self.idx == 0 {
-                self.idx = self.frequency;
-                Err(io::Error::new(io::ErrorKind::WouldBlock, "BlockingRead"))
-            } else {
-                let len = cmp::min(self.idx, buf.len());
-                let bytes_read = try!(self.read.read(&mut buf[..len]));
-                self.idx -= bytes_read;
-                Ok(bytes_read)
-            }
-        }
-    }
-
-    /// Wraps a `Write` instance and introduces blocking.
-    struct BlockingWrite<W> where W: Write {
-        /// The wrapped writer
-        writer: W,
-
-        /// Number of bytes to write before blocking
-        frequency: usize,
-
-        /// Number of bytes written since last blocking
-        idx: usize,
-    }
-
-    impl <W> BlockingWrite<W> where W: Write {
-        fn new(writer: W, frequency: usize) -> BlockingWrite<W> {
-            BlockingWrite { writer: writer, frequency: frequency, idx: 0 }
-        }
-        fn into_writer(self) -> W {
-            self.writer
-        }
-    }
-
-    impl <W> Write for BlockingWrite<W> where W: Write {
-        fn write(&mut self, buf: &[u8]) -> io::Result<usize> {
-            if self.idx == 0 {
-                self.idx = self.frequency;
-                Err(io::Error::new(io::ErrorKind::WouldBlock, "BlockingWrite"))
-            } else {
-                let len = cmp::min(self.idx, buf.len());
-                let bytes_written = try!(self.writer.write(&buf[..len]));
-                self.idx -= bytes_written;
-                Ok(bytes_written)
-            }
-        }
-        fn flush(&mut self) -> io::Result<()> { Ok(()) }
-    }
-
-    #[test]
-    fn check_round_trip_async() {
-        fn round_trip(read_block_frequency: usize,
-                      write_block_frequency: usize,
-                      segments: Vec<Vec<Word>>) -> TestResult
-        {
-            if segments.len() == 0 || read_block_frequency == 0 || write_block_frequency == 0 {
-                return TestResult::discard();
-            }
-
-            let (mut read, segments) = {
-                let cursor = Cursor::new(Vec::new());
-                let writer = BlockingWrite::new(cursor, write_block_frequency);
-                let mut state = write_message_async(writer, segments);
-
-                let mut result = state.poll().unwrap();
-                while let Async::NotReady = result {
-                    result = state.poll().unwrap();
-                }
-
-                match result {
-                    Async::NotReady => unreachable!(),
-                    Async::Ready((writer, m)) => {
-                        let mut cursor = writer.into_writer();
-                        cursor.set_position(0);
-                        (BlockingRead::new(cursor, read_block_frequency), m)
-                    }
-                }
-            };
-
-            let message = {
-                let mut state = read_message_async(&mut read, Default::default());
-                let mut result = state.poll().unwrap();
-                while let Async::NotReady = result {
-                    result = state.poll().unwrap();
-                }
-                match result {
-                    Async::Ready((_, m)) => m,
-                    _ => unreachable!(),
-                }
-            };
-            let message_segments = message.into_segments();
-
-            TestResult::from_bool(segments.iter().enumerate().all(|(i, segment)| {
-                &segment[..] == message_segments.get_segment(i as u32).unwrap()
-            }))
-        }
-
-        quickcheck(round_trip as fn(usize, usize, Vec<Vec<Word>>) -> TestResult);
     }
 }
